chore: improve FormLabel exports
